requirements.txt
google-generativeai>=0.3.2
streamlit>=1.28.0
fpdf
PyPDF2
Pandas